Please review precourse work 2
I have added all the complexities inside document
