--- Exercise_4.py
@@ -1,18 +1,0 @@
-# Python program for implementation of MergeSort 
-def mergeSort(arr):
-  
-  #write your code here
-  
-# Code to print the list 
-def printList(arr): 
-    
-    #write your code here
-  
-# driver code to test the above code 
-if __name__ == '__main__': 
-    arr = [12, 11, 13, 5, 6, 7]  
-    print ("Given array is", end="\n")  
-    printList(arr) 
-    mergeSort(arr) 
-    print("Sorted array is: ", end="\n") 
-    printList(arr)
