--- Exercise_5.py
@@ -1,10 +1,0 @@
-# Python program for implementation of Quicksort
-
-# This function is same in both iterative and recursive
-def partition(arr, l, h):
-  #write your code here
-
-
-def quickSortIterative(arr, l, h):
-  #write your code here
-
