+++ IterativeQuickSort.java
@@ -1,0 +1,73 @@
+import java.util.Stack;
+
+class IterativeQuickSort {
+    void swap(int arr[], int i, int j)
+    {
+        //Try swapping without extra variable
+        int temp = arr[i];
+        arr[i] = arr[j];
+        arr[j] = temp;
+    }
+
+    /* This function is same in both iterative and
+       recursive*/
+    int partition(int arr[], int l, int h)
+    {
+        //Compare elements and swap.
+        int pivot = arr[h];
+        int i = (l - 1);
+
+        for (int j = l; j <= h - 1; j++) {
+            if (arr[j] <= pivot) {
+                i++;
+                swap(arr, i, j);
+            }
+        }
+        swap(arr, i + 1, h);
+
+        return i + 1;
+    }
+
+    // Sorts arr[l..h] using iterative QuickSort
+    void QuickSort(int arr[], int l, int h)
+    {
+        //Try using Stack Data Structure to remove recursion.
+        Stack<Integer> stack = new Stack();
+        stack.push(l);
+        stack.push(h);
+
+        while (!stack.isEmpty()) {
+            int end = stack.pop();
+            int start = stack.pop();
+            int pivot = partition(arr, start, end);
+            if (pivot - 1 > start) {
+                stack.push(start);
+                stack.push(pivot - 1);
+            }
+            if (pivot + 1 < end) {
+                stack.push(pivot + 1);
+                stack.push(end);
+            }
+        }
+    }
+
+    // time complexity = O(n log n)
+    // space complexity = O(n)
+
+    // A utility function to print contents of arr
+    void printArr(int arr[], int n)
+    {
+        int i;
+        for (i = 0; i < n; ++i)
+            System.out.print(arr[i] + " ");
+    }
+
+    // Driver code to test above
+    public static void main(String args[])
+    {
+        IterativeQuickSort ob = new IterativeQuickSort();
+        int arr[] = { 4, 3, 5, 2, 1, 3, 2, 3 };
+        ob.QuickSort(arr, 0, arr.length - 1);
+        ob.printArr(arr, arr.length);
+    }
+}
