--- README.md
@@ -1,15 +1,0 @@
-# PreCourse_2
-
-# All Instructions are already provided in the respective files.
-
-Exercise_1 : Binary Search.
-
-Exercise_2 : Quick sort.
-
-Exercise_3 : Find Mid Point of a Singly Linked List.
-
-Exercise_4 : Merge Sort.
-
-Exercise_5 : Iterative Quick Sort.
-
-*After completing the project kindly submit a pull request*
